Refactored to move some commands to a new service
